Created CLI and ClientSocket handler classes
Splitted the responsabilities of the original ChatClient into two classes: com.comp1459.ClientCLI.ClientCLI.java and com.comp1459.ClientSocket.java

ClientCLI - Its sole purpose is to receive events from the clientSocket and show adequate messages to the user. Provides an interface to the user to interact with the client socket which will communicate with server
ClientSocket - Multiple instances can be created. Its purpose is to create a client socket, communicate with the server and emitting events.